+++ src/LiveComponent/src/Util/FingerprintCalculator.php
@@ -15,8 +15,15 @@ namespace Symfony\UX\LiveComponent\Util;
 
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
+use Symfony\UX\LiveComponent\Metadata\LiveComponentMetadata;
 
 /**
+ * Calculates a fingerprint that is unique to the original input props passed
+ * into the component and that also set props that "accept updates from parent".
+ *
+ * This is used for child components, to determine when the parent has changed
+ * certain props that should trigger the child to re-render.
+ *
  * @author Ryan Weaver <ryan@symfonycasts.com>
  *
  * @experimental
@@ -31,9 +38,15 @@ class FingerprintCalculator
     ) {
     }
 
-    public function calculateFingerprint(array $data): string
+    public function calculateFingerprint(array $inputProps, LiveComponentMetadata $liveMetadata): string
     {
-        $normalizedData = $this->objectNormalizer->normalize($data, context: [LiveComponentHydrator::LIVE_CONTEXT => true]);
+        $fingerprintProps = $liveMetadata->getOnlyPropsThatAcceptUpdatesFromParent($inputProps);
+
+        if (0 === \count($fingerprintProps)) {
+            return '';
+        }
+
+        $normalizedData = $this->objectNormalizer->normalize($fingerprintProps, context: [LiveComponentHydrator::LIVE_CONTEXT => true]);
 
         return base64_encode(hash_hmac('sha256', serialize($normalizedData), $this->secret, true));
     }